PSR-4 Full implemented according current namespace. Could be improved moving some class to sub-namespaces so the file tree would be more organized.
