+++ src/Connection.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class Connection
+ * @package Connect
+ */
+class Connection
+{
+    var $id;
+    var $name;
+
+    /**
+     * @var Provider
+     */
+    var $provider;
+
+    /**
+     * @var Vendor
+     */
+    var $vendor;
+}
