+++ src/Contact.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class Contact
+ * @package Connect
+ */
+class Contact
+{
+    var $email;
+    var $first_name;
+    var $last_name;
+
+    /**
+     * @var PhoneNumber
+     */
+    var $phone_number;
+}
