+++ src/ContactInfo.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class ContactInfo
+ * @package Connect
+ */
+class ContactInfo
+{
+    var $address_line1;
+    var $address_line2;
+    var $city;
+    
+    /**
+     * @var Contact
+     */
+    var $contact;
+    var $country;
+    var $postal_code;
+    var $state;
+}
+
