+++ src/Item.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class Item
+ * @package Connect
+ */
+class Item
+{
+    var $id;
+    var $mpn;
+
+    /**
+     * @var int
+     */
+    var $quantity;
+
+    /**
+     * @var int
+     */
+    var $old_quantity;
+}
