+++ src/Logger.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of the Ingram Micro Cloud Blue Connect SDK.
  *
@@ -6,154 +7,6 @@
  */
 
 namespace Connect;
-
-/**
- * Interface for Logger
- *
- * @package Connect
- */
-interface LoggerInterface
-{
-	const LEVEL_TRACE = 4;
-	const LEVEL_DEBUG = 3;
-	const LEVEL_INFO  = 2;
-	const LEVEL_ERROR = 1;
-	const LEVEL_FATAL = 0;
-
-	const LEVELS = array(
-			self::LEVEL_FATAL => 'FATAL',
-			self::LEVEL_ERROR => 'ERROR',
-			self::LEVEL_INFO  => 'INFO',
-			self::LEVEL_DEBUG => 'DEBUG',
-			self::LEVEL_TRACE => 'TRACE'
-	);
-
-    /**
-     * Logs trace messages
-     *
-     * @param string $message
-     */
-    public function trace($message);
-
-    /**
-     * Logs debug messages
-     *
-     * @param string $message
-     */
-    public function debug($message);
-
-    /**
-     * Logs info messages
-     *
-     * @param string $message
-     */
-    public function info($message);
-
-    /**
-     * Logs error messages
-     *
-     * @param string $message
-     */
-    public function error($message);
-
-    /**
-     * Logs fatal messages
-     *
-     * @param string $message
-     */
-    public function fatal($message);
-
-    /**
-     * Logs message of any level
-     *
-     * @param int    $level   - one of LEVEL_* constants
-     * @param string $message - message to log
-     */
-    public function log($level, $message);
-
-    /**
-     * Write log record into log
-     *
-     * @param LogRecord $record
-     */
-    public function write($record);
-
-    /**
-     * Set logger logLevel (one of LEVEL_* constants)
-     * @param int $level
-     */
-    public function setLogLevel($level);
-
-    /**
-     * Dump session log to current log, and reset session
-     */
-    public function dump();
-}
-
-/**
- * LogRecord class, instance created for every logging action
- *
- * @package Connect
- */
-class LogRecord
-{
-	public $level;
-	public $message;
-	public $time;
-
-	public function __construct($level, $message, $time = 0)
-	{
-		$this->level = $level;
-		$this->message = $message;
-		$this->time = ($time === 0) ? time() : $time;
-	}
-}
-
-/**
- * Collects all log records for one processing batch
- * In case of exception these records will be added to log for analysis
- *
- * @package Connect
- */
-class LogSession
-{
-	private $sessionLog = array();
-
-	/**
-	 * Add logging record to session log
-	 *
-	 * @param LogRecord $record
-	 */
-	public function addRecord($record)
-	{
-		$this->sessionLog[] = $record;
-	}
-
-	/**
-	 * Reset session log
-	 */
-	public function reset()
-	{
-		$this->sessionLog = array();
-	}
-
-	/**
-	 * Dump session log to specified logger
-	 *
-	 * @param LoggerInterface $log
-	 */
-	public function dumpTo($log)
-	{
-		if (!$this->sessionLog)
-			return;
-
-		$log->write(new LogRecord(null, "=== Detailed session log dump begin ==================", null));
-		foreach($this->sessionLog as $r) {
-			$log->write($r); // historical log dumped with actual event times
-		}
-		$log->write(new LogRecord(null, "=== Detailed session log dump end ====================", null));
-	}
-}
 
  /**
  * Default Logger for Cloud Blue Connect SDK.
