+++ src/PhoneNumber.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class PhoneNumber
+ * @package Connect
+ */
+class PhoneNumber
+{
+    var $country_code;
+    var $area_code;
+    var $phone_number;
+    var $extension;
+}
