+++ src/Provider.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class Provider
+ * @package Connect
+ */
+class Provider
+{
+    var $id;
+    var $name;
+}
