+++ src/Tier.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * This file is part of the Ingram Micro Cloud Blue Connect SDK.
+ *
+ * @copyright (c) 2018. Ingram Micro. All Rights Reserved.
+ */
+
+namespace Connect;
+
+/**
+ * Class Tier
+ * @package Connect
+ */
+class Tier
+{
+    var $id;
+    var $external_id;
+    var $name;
+
+    /**
+     * @var ContactInfo
+     */
+    var $contact_info;
+}
